requirements.txt
django
south
django-crispy-forms
django-cryptocoin